Defined the set of golden numbers.
All that remains is to prove that the set is countbly infinte by proving there exists a bijection from the set to the set of natural numbers. How do I prove this bijection? Do i even need to porve this bijection since the set is defined as a function of n a nonzero natural number? Also can I expand this and turn this into a thesis? And how do i graph the spirals?